#9 Fixed issue with overflowing Nav bar
Renamed 'Layout' class to 'BottomNavBar' for easier identification.

Re-did bit of how the nav bar was implemented. The nav bar func/class now just returns a Paper component with the nav bar inside rather than divs/ containers, etc. The issue with the overflow and instances where the positioning was iffy now seem to be fixed!